modularize arrivals page
