+++ src/features/track/track.module.css
@@ -1,0 +1,41 @@
+.dropdownLabel {
+  font-size: 12px;
+  display: block;
+  margin-bottom: 4px;
+}
+
+.backButton {
+  display: inline;
+  align-items: center;
+  justify-content: flex-start;
+  padding: 8px 16px 8px 0px;
+}
+
+.backLink {
+  text-decoration: none;
+  display: block;
+  width: fit-content;
+}
+
+.titleContainer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.fullWidth {
+  width: 100%;
+}
+
+.alignLeft {
+  text-align: left;
+}
+
+.errorTitle {
+  font-size: 24px;
+}
+
+.warningIcon {
+  font-size: 48px;
+  color: #faad14;
+}
